As we said on the last one of these, fc2dab717:

  This setting is not to be confused with the targetSdkVersion. The
  latter goes into the built manifest, and affects a wide range of
  behavior, so bumping it requires careful testing. This is used
  purely at build time, and should have no effect on runtime
  behavior.

This value needs to be at least as high as targetSdkVersion, and we need to bump that to 34 soon (#5877). So, do this to prepare for that.

No build failures or warnings appeared for me with
  tools/test native --all-files
.

Related: #5877