@gnprice 2024-07-10
> No build failures or warnings appeared for me with
> tools/test native --all-files

So unfortunately we have a large existing base of Android build warnings, mostly (though not entirely) caused by our dependencies. `tools/test native` therefore suppresses them:

```bash
    tools/gradle -q :app:assembleDebug :app:assembleDebugUnitTest \
        :app:bundleDebug || return

    # The `-q` suppresses noise from warnings about obsolete build config
    # in our dependencies from the React Native ecosystem.
    # But it also suppresses the names of tests that failed.
    # So on failure, rerun without it.
    tools/gradle -q :app:testDebugUnitTest \
        || tools/gradle :app:testDebugUnitTest
```

(This is a situation I hope to avoid getting into with zulip-flutter — zulip/zulip-flutter#796 causes CI to enforce that there are zero Kotlin compiler warnings, at least, and hopefully we can maintain that.)

So to see the warnings, use `tools/gradle :app:assembleDebug` with no `-q`.

You may have to add `--rerun-tasks` to cause Gradle to rebuild things it's previously built, and therefore report warnings there again.

I think probably the right way to check for warnings here is:
* Run the build at main; save the output to a file like `tools/gradle … >baseline.log 2>&1`.
* Run the build after the change, similarly saving it.
* Use `git diff --no-index` to look for changes.
  * And hope that the order the log lines appear in is consistent. If it isn't and the diff therefore has a lot of noise… try running `sort` on each file and comparing the results. If they match, then great; where they don't match, it might be a bit annoying to back out what's going on because neighboring lines will have been scattered by the sort, but at least it gives a comprehensive list of things that need figuring out.